+++ gatherer/gatherers.go
@@ -1,8 +1,7 @@
-package run
+package gatherer
 
 import (
 	c "github.com/cloudurable/metricsd/common"
-	g "github.com/cloudurable/metricsd/gatherer"
 )
 
 func LoadGatherers(config *c.Config) ([]c.MetricsGatherer) {
@@ -12,25 +11,25 @@ func LoadGatherers(config *c.Config) ([]c.MetricsGatherer) {
 	for _,provider := range config.Gatherers {
 		switch provider {
 		case c.PROVIDER_CPU:
-			cpu := g.NewCPUMetricsGatherer(nil, config)
+			cpu := NewCPUMetricsGatherer(nil, config)
 			if cpu != nil {
 				gatherers = append(gatherers, cpu)
 			}
 
 		case c.PROVIDER_DISK:
-			disk := g.NewDiskMetricsGatherer(nil, config)
+			disk := NewDiskMetricsGatherer(nil, config)
 			if disk != nil {
 				gatherers = append(gatherers, disk)
 			}
 
 		case c.PROVIDER_FREE:
-			free := g.NewFreeMetricGatherer(nil, config)
+			free := NewFreeMetricGatherer(nil, config)
 			if free != nil {
 				gatherers = append(gatherers, free)
 			}
 
 		case c.PROVIDER_NODETOOL:
-			for _, nodetool := range g.NewNodetoolMetricGatherers(nil, config) {
+			for _, nodetool := range NewNodetoolMetricGatherers(nil, config) {
 				gatherers = append(gatherers, nodetool)
 			}
 		}
