+++ repeater/repeaters.go
@@ -1,8 +1,7 @@
-package run
+package repeater
 
 import (
 	c "github.com/cloudurable/metricsd/common"
-    r "github.com/cloudurable/metricsd/repeater"
 )
 
 func LoadRepeaters(config *c.Config) ([]c.MetricsRepeater) {
@@ -12,15 +11,15 @@ func LoadRepeaters(config *c.Config) ([]c.MetricsRepeater) {
 	for _,provider := range config.Repeaters {
 		switch provider {
 		case c.REPEATER_AWS:
-			aws := r.NewAwsCloudMetricRepeater(config);
-			if aws != nil {
-				repeaters = append(repeaters, aws)
+			repeater := NewAwsCloudMetricRepeater(config);
+			if repeater != nil {
+				repeaters = append(repeaters, repeater)
 			}
 
 		case c.REPEATER_LOGGER:
-			lgr := r.NewLogMetricsRepeater();
-			if lgr != nil {
-				repeaters = append(repeaters, lgr)
+			repeater := NewLogMetricsRepeater();
+			if repeater != nil {
+				repeaters = append(repeaters, repeater)
 			}
 		}
 	}
